devel/p5-Mock-Quick: import p5-Mock-Quick-1.111
Mock-Quick is here to solve the current problems with Mocking
libraries.

There are a couple Mocking libraries available on CPAN. The primary
problems with these libraries include verbose syntax, and most
importantly side-effects. Some Mocking libraries expect you to mock
a specific class, and will unload it then redefine it. This is
particularly a problem if you only want to override a class on a
lexical level.

Mock-Quick provides a declarative mocking interface that results
in a very concise, but clear syntax. There are separate facilities
for mocking object instances, and classes. You can quickly create
an instance of an object with custom attributes and methods. You
can also quickly create an anonymous class, optionally inheriting
from another, with whatever methods you desire.

Mock-Quick also provides a tool that provides an OO interface to
overriding methods in existing classes. This tool also allows for
the restoration of the original class methods. Best of all this is
a localized tool, when your control object falls out of scope the
original class is restored.